moved App component and supporting files to directory with index
