--- src/components/App.module.scss
@@ -1,26 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  width: 200px;
-}
-
-.App-header {
-  background-color: #47d7ac;
-  padding: 20px;
-  color: white;
-}
-
-.App-title {
-  font-size: 1.5em;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-.App-content {
-  max-width: 1280px;
-  margin: auto;
-}
